+++ R/barycentricProjection.R
@@ -51,7 +51,7 @@
 #' pp <- 6
 #' overlap <- "low"
 #' design <- "A"
-#' estimate <- "ATE"
+#' estimate <- "ATT"
 #' power <- 2
 #' data <- causalOT::Hainmueller$new(n = n, p = pp,
 #' design = design, overlap = overlap)
@@ -65,8 +65,10 @@
 #'   
 #'  df <- data.frame(y = data$get_y(), z = data$get_z(), data$get_x())
 #'   
-#'  fit <- causalOT::barycentric_projection(y ~ ., data = df, weight = weights,
-#'     separate.samples.on = "z")
+#'  fit <- causalOT::barycentric_projection(y ~ ., data = df, 
+#'     weight = weights,
+#'     separate.samples.on = "z",
+#'     niter = 2)
 #'  inherits(fit, "bp")
 #'  }
 barycentric_projection <- function(formula, data, weights, 
@@ -172,7 +174,7 @@ barycentric_projection <- function(formula, data, weights,
 #' pp <- 6
 #' overlap <- "low"
 #' design <- "A"
-#' estimate <- "ATE"
+#' estimate <- "ATT"
 #' power <- 2
 #' data <- causalOT::Hainmueller$new(n = n, p = pp,
 #' design = design, overlap = overlap)
@@ -189,18 +191,19 @@ barycentric_projection <- function(formula, data, weights,
 #'  # undebiased
 #'  fit <- causalOT::barycentric_projection(y ~ ., data = df, 
 #'     weight = weights,
-#'     separate.samples.on = "z")
-#'  
+#'     separate.samples.on = "z", niter = 2)
+#'     
+#'  #debiased
 #'  fit_d <- causalOT::barycentric_projection(y ~ ., data = df, 
 #'     weight = weights,
-#'     separate.samples.on = "z", debias = TRUE)
+#'     separate.samples.on = "z", debias = TRUE, niter = 2)
 #'  
 #'  # predictions, without new data
 #'  undebiased_predictions <- predict(fit,   source.sample = df$z)
 #'  debiased_predictions   <- predict(fit_d, source.sample = df$z)
 #'  
-#'  all.equal(undebiased_predictions, df$y) # FALSE
-#'  all.equal(debiased_predictions, df$y) # TRUE
+#'  isTRUE(all.equal(unname(undebiased_predictions), df$y)) # FALSE
+#'  isTRUE(all.equal(unname(debiased_predictions), df$y)) # TRUE
 #'  }
 predict.bp <- function(object, newdata = NULL, source.sample, cost_function = NULL, niter = 1e3, tol = 1e-7, ...) {