+++ R/weightsClass.R
@@ -11,6 +11,7 @@
 #' @slot call The call used to construct the weights.
 #'
 #' @docType class
+#' @include dataHolder.R
 #' 
 #' @details This object is returned by the `calc_weight` function in this package. The slots can be accessed as any S4 object. There is no publicly accessible constructor function. 
 #' 
@@ -27,37 +28,6 @@ setClass("causalWeights", slots = c(w0 = "numeric", w1 = "numeric",
 
 setGeneric("causalWeights", function(object1, object2, ...) standardGeneric("causalWeights"))
 
-setMethod("causalWeights", signature(object1 = "gridSearch", object2 = "numeric"), 
-function(object1, object2, ...) {
-  penalty <- list(...)[[1]]
-  estimand <- object1@estimand
-  
-  if (estimand == "ATE.T" || estimand == "ATC") {
-    w0 <- object1@solver$b
-    w1 <- as.numeric(object2)
-  } else if (estimand == "ATE.C" || estimand == "ATT") {
-    w0 <- as.numeric(object2)
-    w1 <- object1@solver$b
-  } else {
-    stop("estimand not found!")
-  }
-  if(!is.list(penalty$penalty)) penalty$penalty <- as.list(penalty$penalty)
-  
-  methods::new("causalWeights",
-      w0 = w0,
-      w1 = w1,
-      estimand = object1@estimand,
-      method  = object1@method,
-      penalty = penalty$penalty,
-      info    = list(metric = penalty$metric,
-                     penalty.grid = penalty$penalty.grid,
-                     gradients = NULL,
-                     hessian = NULL
-                     ),
-      data = dataHolder(x = matrix(0,0,0), z = numeric(0)),
-      call = call("calc_weight"))
-}
-)
 
 setMethod("causalWeights", signature(object1 = "causalWeights", object2 = "causalWeights"), 
 function(object1, object2, ...) {