+++ inst/stan/deprecated/include/stan_meta_header.hpp
@@ -1,2 +1,0 @@
-// Insert all #include<foo.hpp> statements here
-// #include <RcppParallel.h>